refactor: unify mock git client implementations
- Move cmd/test_utils.go to internal/testutil/git_client.go
- Remove duplicate MockGitClient implementations from config/config_test.go and cmd/interactive_test.go
- Update all test files to use unified testutil.NewMockGitClient()
- Resolve import cycle issues by centralizing mock in internal/testutil
- Improve test maintainability with single source of truth for mocks
- All tests pass with unified mock implementation
