+++ config/config_test.go
@@ -8,121 +8,15 @@ import (
 	"strings"
 	"testing"
 
+	"github.com/bmf-san/ggc/v4/internal/testutil"
 	"gopkg.in/yaml.v3"
 )
 
-// MockGitClient implements git.Clienter for testing (only config operations)
-type MockGitClient struct {
-	configs map[string]string
-	setErr  error
-	getErr  error
-}
 
-// NewMockGitClient creates a new mock git client
-func NewMockGitClient() *MockGitClient {
-	return &MockGitClient{
-		configs: make(map[string]string),
-	}
-}
-
-// ConfigSetGlobal mocks setting global git config
-func (m *MockGitClient) ConfigSetGlobal(key, value string) error {
-	if m.setErr != nil {
-		return m.setErr
-	}
-	m.configs[key] = value
-	return nil
-}
-
-// ConfigGetGlobal mocks getting global git config
-func (m *MockGitClient) ConfigGetGlobal(key string) (string, error) {
-	if m.getErr != nil {
-		return "", m.getErr
-	}
-	return m.configs[key], nil
-}
-
-// Implement other required methods from git.Clienter interface (minimal stubs)
-func (m *MockGitClient) GetCurrentBranch() (string, error)     { return "main", nil }
-func (m *MockGitClient) GetBranchName() (string, error)        { return "main", nil }
-func (m *MockGitClient) GetGitStatus() (string, error)         { return "", nil }
-func (m *MockGitClient) Status() (string, error)               { return "", nil }
-func (m *MockGitClient) StatusShort() (string, error)          { return "", nil }
-func (m *MockGitClient) StatusWithColor() (string, error)      { return "", nil }
-func (m *MockGitClient) StatusShortWithColor() (string, error) { return "", nil }
-func (m *MockGitClient) Add(_ ...string) error                 { return nil }
-func (m *MockGitClient) AddInteractive() error                 { return nil }
-func (m *MockGitClient) Commit(_ string) error                 { return nil }
-func (m *MockGitClient) CommitAmend() error                    { return nil }
-func (m *MockGitClient) CommitAmendNoEdit() error              { return nil }
-func (m *MockGitClient) CommitAmendWithMessage(_ string) error { return nil }
-func (m *MockGitClient) CommitAllowEmpty() error               { return nil }
-func (m *MockGitClient) Diff() (string, error)                 { return "", nil }
-func (m *MockGitClient) DiffStaged() (string, error)           { return "", nil }
-func (m *MockGitClient) DiffHead() (string, error)             { return "", nil }
-func (m *MockGitClient) ListLocalBranches() ([]string, error)  { return nil, nil }
-func (m *MockGitClient) ListRemoteBranches() ([]string, error) { return nil, nil }
-func (m *MockGitClient) CheckoutNewBranch(_ string) error      { return nil }
-func (m *MockGitClient) CheckoutBranch(_ string) error         { return nil }
-func (m *MockGitClient) CheckoutNewBranchFromRemote(_, _ string) error {
-	return nil
-}
-func (m *MockGitClient) DeleteBranch(_ string) error                     { return nil }
-func (m *MockGitClient) ListMergedBranches() ([]string, error)           { return nil, nil }
-func (m *MockGitClient) RevParseVerify(_ string) bool                    { return false }
-func (m *MockGitClient) Push(_ bool) error                               { return nil }
-func (m *MockGitClient) Pull(_ bool) error                               { return nil }
-func (m *MockGitClient) Fetch(_ bool) error                              { return nil }
-func (m *MockGitClient) RemoteList() error                               { return nil }
-func (m *MockGitClient) RemoteAdd(_, _ string) error                     { return nil }
-func (m *MockGitClient) RemoteRemove(_ string) error                     { return nil }
-func (m *MockGitClient) RemoteSetURL(_, _ string) error                  { return nil }
-func (m *MockGitClient) TagList(_ []string) error                        { return nil }
-func (m *MockGitClient) TagCreate(_, _ string) error                     { return nil }
-func (m *MockGitClient) TagCreateAnnotated(_, _ string) error            { return nil }
-func (m *MockGitClient) TagDelete(_ []string) error                      { return nil }
-func (m *MockGitClient) TagPush(_, _ string) error                       { return nil }
-func (m *MockGitClient) TagPushAll(_ string) error                       { return nil }
-func (m *MockGitClient) TagShow(_ string) error                          { return nil }
-func (m *MockGitClient) GetLatestTag() (string, error)                   { return "", nil }
-func (m *MockGitClient) TagExists(_ string) bool                         { return false }
-func (m *MockGitClient) GetTagCommit(_ string) (string, error)           { return "", nil }
-func (m *MockGitClient) LogSimple() error                                { return nil }
-func (m *MockGitClient) LogGraph() error                                 { return nil }
-func (m *MockGitClient) LogOneline(_, _ string) (string, error)          { return "", nil }
-func (m *MockGitClient) RebaseInteractive(_ int) error                   { return nil }
-func (m *MockGitClient) GetUpstreamBranch(_ string) (string, error)      { return "", nil }
-func (m *MockGitClient) Stash() error                                    { return nil }
-func (m *MockGitClient) StashList() (string, error)                      { return "", nil }
-func (m *MockGitClient) StashShow(_ string) error                        { return nil }
-func (m *MockGitClient) StashApply(_ string) error                       { return nil }
-func (m *MockGitClient) StashPop(_ string) error                         { return nil }
-func (m *MockGitClient) StashDrop(_ string) error                        { return nil }
-func (m *MockGitClient) StashClear() error                               { return nil }
-func (m *MockGitClient) RestoreWorkingDir(_ ...string) error             { return nil }
-func (m *MockGitClient) RestoreStaged(_ ...string) error                 { return nil }
-func (m *MockGitClient) RestoreFromCommit(_ string, _ ...string) error   { return nil }
-func (m *MockGitClient) RestoreAll() error                               { return nil }
-func (m *MockGitClient) RestoreAllStaged() error                         { return nil }
-func (m *MockGitClient) ConfigGet(_ string) (string, error)              { return "", nil }
-func (m *MockGitClient) ConfigSet(_, _ string) error                     { return nil }
-func (m *MockGitClient) ResetHardAndClean() error                        { return nil }
-func (m *MockGitClient) ResetHard(_ string) error                        { return nil }
-func (m *MockGitClient) CleanFiles() error                               { return nil }
-func (m *MockGitClient) CleanDirectories() error                         { return nil }
-func (m *MockGitClient) CleanDirs() error                                { return nil }
-func (m *MockGitClient) CleanDryRun() (string, error)                    { return "", nil }
-func (m *MockGitClient) CleanFilesForce(_ []string) error                { return nil }
-func (m *MockGitClient) CleanInteractive() error                         { return nil }
-func (m *MockGitClient) ListFiles() (string, error)                      { return "", nil }
-func (m *MockGitClient) GetUpstreamBranchName(_ string) (string, error)  { return "", nil }
-func (m *MockGitClient) GetAheadBehindCount(_, _ string) (string, error) { return "", nil }
-func (m *MockGitClient) GetVersion() (string, error)                     { return "test-version", nil }
-func (m *MockGitClient) GetCommitHash() (string, error)                  { return "test-commit", nil }
 
 // newTestConfigManager creates a config manager for testing without executing git commands
 func newTestConfigManager() *Manager {
-	mockClient := NewMockGitClient()
+	mockClient := testutil.NewMockGitClient()
 	config := getDefaultConfig(mockClient)
 
 	return &Manager{
@@ -134,7 +28,7 @@ func newTestConfigManager() *Manager {
 // TestGetDefaultConfig tests the default configuration values
 func TestGetDefaultConfig(t *testing.T) {
 	// Use mock client to avoid executing real git commands
-	mockClient := NewMockGitClient()
+	mockClient := testutil.NewMockGitClient()
 	config := getDefaultConfig(mockClient)
 
 	// test default values
@@ -177,7 +71,7 @@ func TestGetDefaultConfig(t *testing.T) {
 
 // TestNewConfigManager tests the creation of a new config manager
 func TestNewConfigManager(t *testing.T) {
-	mockClient := NewMockGitClient()
+	mockClient := testutil.NewMockGitClient()
 	cm := NewConfigManager(mockClient)
 
 	if cm == nil {
@@ -251,9 +145,9 @@ integration:
 		t.Fatalf("Failed to write test config: %v", err)
 	}
 
-	cm := NewConfigManager(NewMockGitClient())
+	cm := NewConfigManager(testutil.NewMockGitClient())
 	// Use mock git client to avoid real git config operations
-	cm.gitClient = NewMockGitClient()
+	cm.gitClient = testutil.NewMockGitClient()
 	err = cm.loadFromFile(configPath)
 	if err != nil {
 		t.Fatalf("Failed to load config from file: %v", err)
@@ -281,9 +175,9 @@ integration:
 
 // TestLoad tests the Load method with no config file
 func TestLoad(t *testing.T) {
-	cm := NewConfigManager(NewMockGitClient())
+	cm := NewConfigManager(testutil.NewMockGitClient())
 	// Use mock git client to avoid real git config operations
-	cm.gitClient = NewMockGitClient()
+	cm.gitClient = testutil.NewMockGitClient()
 
 	originalHome := os.Getenv("HOME")
 	tempDir := t.TempDir()
@@ -312,10 +206,10 @@ func TestSave(t *testing.T) {
 	tempDir := t.TempDir()
 	configPath := filepath.Join(tempDir, "test-save.yaml")
 
-	cm := NewConfigManager(NewMockGitClient())
+	cm := NewConfigManager(testutil.NewMockGitClient())
 	cm.configPath = configPath
 	// Use mock git client to avoid real git config operations
-	cm.gitClient = NewMockGitClient()
+	cm.gitClient = testutil.NewMockGitClient()
 
 	cm.config.Default.Branch = "development"
 	cm.config.UI.Color = false
@@ -382,7 +276,7 @@ func TestGetValueByPath(t *testing.T) {
 
 // TestGetValueByPathErrors tests error cases for getValueByPath
 func TestGetValueByPathErrors(t *testing.T) {
-	cm := NewConfigManager(NewMockGitClient())
+	cm := NewConfigManager(testutil.NewMockGitClient())
 
 	testCases := []string{
 		"nonexistent.field",
@@ -453,10 +347,10 @@ func TestSet(t *testing.T) {
 	tempDir := t.TempDir()
 	configPath := filepath.Join(tempDir, "test-set.yaml")
 
-	cm := NewConfigManager(NewMockGitClient())
+	cm := NewConfigManager(testutil.NewMockGitClient())
 	cm.configPath = configPath
 	// Use mock git client to avoid real git config operations
-	cm.gitClient = NewMockGitClient()
+	cm.gitClient = testutil.NewMockGitClient()
 
 	err := cm.Set("default.branch", "develop")
 	if err != nil {
@@ -595,9 +489,9 @@ func TestLoadConfig(t *testing.T) {
 		}
 	}()
 
-	cm := NewConfigManager(NewMockGitClient())
+	cm := NewConfigManager(testutil.NewMockGitClient())
 	// Use mock git client to avoid real git config operations
-	cm.gitClient = NewMockGitClient()
+	cm.gitClient = testutil.NewMockGitClient()
 	cm.LoadConfig()
 
 	if cm.config == nil {
@@ -652,9 +546,9 @@ ui:
 		t.Fatalf("Failed to write invalid config: %v", err)
 	}
 
-	cm := NewConfigManager(NewMockGitClient())
+	cm := NewConfigManager(testutil.NewMockGitClient())
 	// Use mock git client to avoid real git config operations
-	cm.gitClient = NewMockGitClient()
+	cm.gitClient = testutil.NewMockGitClient()
 	err = cm.loadFromFile(configPath)
 	if err == nil {
 		t.Error("Expected error when loading invalid YAML")
@@ -1114,9 +1008,9 @@ func stringifyValue(val any) string {
 // TestManagerLoadConfig tests the LoadConfig method error paths
 func TestManagerLoadConfig(t *testing.T) {
 	t.Run("LoadConfig with invalid path executes without panic", func(t *testing.T) {
-		cm := NewConfigManager(NewMockGitClient())
+		cm := NewConfigManager(testutil.NewMockGitClient())
 		// Use mock git client to avoid real git config operations
-		cm.gitClient = NewMockGitClient()
+		cm.gitClient = testutil.NewMockGitClient()
 		cm.configPath = "/nonexistent/directory/config.yaml"
 
 		// This should not panic, even if file operations fail
@@ -1136,9 +1030,9 @@ func TestManagerLoadConfig(t *testing.T) {
 	})
 
 	t.Run("LoadConfig handles missing directory gracefully", func(t *testing.T) {
-		cm := NewConfigManager(NewMockGitClient())
+		cm := NewConfigManager(testutil.NewMockGitClient())
 		// Use mock git client to avoid real git config operations
-		cm.gitClient = NewMockGitClient()
+		cm.gitClient = testutil.NewMockGitClient()
 		cm.configPath = "/definitely/nonexistent/path/config.yaml"
 
 		// Should not panic or crash
